Various bug fixes for semantic and syntax tests.

+++ test/libsolidity/semanticTests/storage/packed_storage_signed.sol
@@ -6,16 +6,16 @@ contract C {
 
     function test()
         public
-        returns (uint x1, uint x2, uint x3, uint x4)
+        returns (uint256 x1, uint256 x2, uint256 x3, uint256 x4)
     {
         a = -2;
         unchecked {
             b = (0 - uint8(a)) * 2;
             c = a * int8(120) * int8(121);
         }
-        x1 = uint(int(a));
+        x1 = uint256(int256(a));
         x2 = b;
-        x3 = uint(int(c));
+        x3 = uint256(int256(c));
         x4 = d;
     }
 }
@@ -23,4 +23,4 @@ contract C {
 // ====
 // compileViaYul: also
 // ----
-// test() ->
+// test() -> FAILURE, 4
